Add initial working example
This commit adds the first draft of a GreaseMonkey script that is able
to locate and upgrade one of the input text areas in the PE Console UI.
This version of the script finds `<textarea>` inputs that are created
for adding new class parameters, populates them with a hello world
message, and then upgrades them to CodeMirror editors.

CodeMirror 5 is used as this version is available from CDNs.
CodeMirror 6 is the latest major version, but went all-in on a
"you have to compile it yourself for your usecase" delivery
model.

The PE Console UI is a full-on application that adds most of the
UI components after the page loads. This sort of activity can
be tricky to catch with a user-script, use of a MutationObserver
adapted from:

  https://javascript.info/mutation-observer

The MutationObserver is watching the entire `document.body` element,
so it is likely woefully inefficient. But, this is for demonstration
purposes only, right?