Implement class parameter editing
This commit updates the user script to spawn CodeMirror editors
when class parameters are edited in the PE Console. The script
now handles both adding of new parameters and editing of existing
ones.

CodeMirror editors are now configured to sync back to the underlying
`<textarea>` elements via their own "change" events. This eliminates
the need to hook various buttons in the PE UI.